refactor: share assign drawer subcomponent & hook
Move the `AssignmentCounts` subcomponent of `ActivityAssignDrawer` to a
shared location to be eventually consumed by `ActivityUnassignDrawer`.
Also make its banners hook generic, move it to the shared hooks folder,
and rename it to `useModalBanners`.
